+++ CHANGELOG.md
@@ -94,6 +94,7 @@ There are no breaking API changes in this release, nor any changes to how cluste
 - [PR #2585](https://github.com/rqlite/rqlite/pull/2585), [PR #2586](https://github.com/rqlite/rqlite/pull/2586): Cluster client always retries at least once with new connection.  Fixes issue [#2579](https://github.com/rqlite/rqlite/issues/2579).
 - [PR #2587](https://github.com/rqlite/rqlite/pull/2587): Improve DB-layer expvar stats.
 - [PR #2588](https://github.com/rqlite/rqlite/pull/2588): Improve Snapshot Store expvar stats.
+- [PR #2590](https://github.com/rqlite/rqlite/pull/2590): Improve Snapshot reaping expvar stats.
 
 ## v9.4.5 (March 9th 2026)
 ### Implementation changes and bug fixes